+++ src/js/assets.js
@@ -172,6 +172,27 @@ const isDiffUpdatableAsset = content => {
         data.diffPath.startsWith('%') === false;
 };
 
+const assetKeyFromDiffName = diffName => {
+    const entry = assetCacheRegistry[diffName];
+    if ( entry instanceof Object ) {
+        if ( typeof entry.diffPath === 'string' ) {
+            if ( entry.diffPath.endsWith(`#${diffName}`) ) { return diffName; }
+        }
+        if ( typeof entry.diffName === 'string' ) {
+            if ( entry.diffName === diffName ) { return diffName; }
+        }
+    }
+    for ( const [ assetKey, entry ] of Object.entries(assetCacheRegistry) ) {
+        if ( typeof entry.diffPath === 'string' ) {
+            if ( entry.diffPath.endsWith(`#${diffName}`) ) { return assetKey; }
+        }
+        if ( typeof entry.diffName === 'string' ) {
+            if ( entry.diffName === diffName ) { return assetKey; }
+        }
+    }
+    return '';
+};
+
 /******************************************************************************/
 
 // favorLocal: avoid making network requests whenever possible
@@ -1172,10 +1193,14 @@ const getAssetDiffDetails = assetKey => {
     const out = { name: assetKey };
     const cacheEntry = assetCacheRegistry[assetKey];
     if ( cacheEntry === undefined ) { return; }
-    if ( cacheEntry.diffPath === undefined ) { return; }
-    if ( cacheEntry.diffName === undefined ) { return; }
-    out.diffName = cacheEntry.diffName;
     out.patchPath = cacheEntry.diffPath;
+    if ( out.patchPath === undefined ) { return; }
+    out.diffName = cacheEntry.diffName;
+    if ( out.diffName === undefined ) {
+        const match = /#.+$/.exec(out.patchPath);
+        if ( match === null ) { return; }
+        out.diffName = match[0].slice(1);
+    }
     out.diffExpires = getUpdateAfterTime(assetKey, true);
     out.lastModified = cacheEntry.lastModified;
     const assetEntry = assetSourceRegistry[assetKey];
@@ -1196,8 +1221,6 @@ async function diffUpdater() {
         const assetKey = toUpdate.shift();
         const assetDetails = getAssetDiffDetails(assetKey);
         if ( assetDetails === undefined ) { continue; }
-        if ( assetDetails.patchPath === undefined ) { continue; }
-        if ( assetDetails.diffName === undefined ) { continue; }
         assetDetails.what = 'update';
         if ( (getWriteTime(assetKey) + assetDetails.diffExpires) > now ) {
             assetDetails.fetch = false;
@@ -1219,6 +1242,16 @@ async function diffUpdater() {
             if ( typeof error !== 'string' ) { return; }
             ubolog(`Diff updater: terminate because ${error}`);
         };
+        const checkAndCorrectDiffPath = data => {
+            if ( typeof data.text !== 'string' ) { return; }
+            if ( data.text === '' ) { return; }
+            const assetKey = assetKeyFromDiffName(data.name);
+            if ( assetKey === '' ) { return; }
+            const metadata = extractMetadataFromList(data.text, [ 'Diff-Path' ]);
+            if ( metadata instanceof Object === false ) { return; }
+            if ( metadata.diffPath === data.patchPath ) { return; }
+            assetCacheSetDetails(assetKey, metadata.diffPath);
+        };
         bc.onmessage = ev => {
             const data = ev.data;
             if ( data.what === 'ready' ) {
@@ -1237,9 +1270,11 @@ async function diffUpdater() {
             }
             if ( data.status === 'needtext' ) {
                 ubolog('Diff updater: need text for', data.name);
-                assetCacheRead(data.name).then(result => {
+                const assetKey = assetKeyFromDiffName(data.name);
+                assetCacheRead(assetKey).then(result => {
                     data.text = result.content;
                     data.status = undefined;
+                    checkAndCorrectDiffPath(data);
                     bc.postMessage(data);
                 });
                 return;
@@ -1253,22 +1288,24 @@ async function diffUpdater() {
                     'Diff-Path',
                     'Diff-Expires',
                 ]);
-                assetCacheWrite(data.name, {
+                const assetKey = assetKeyFromDiffName(data.name);
+                assetCacheWrite(assetKey, {
                     content: data.text,
                     resourceTime: metadata.lastModified || 0,
                 });
-                assetCacheSetDetails(data.name, metadata);
-                updaterUpdated.push(data.name);
+                assetCacheSetDetails(assetKey, metadata);
+                updaterUpdated.push(assetKey);
             } else if ( data.error ) {
-                ubolog(`Diff updater: failed to update ${data.name} using ${data.patchPath}, reason: ${data.error}`);
+                ubolog(`Diff updater: failed to update ${data.name} using ${data.patchPath}\n\treason: ${data.error}`);
             } else if ( data.status === 'nopatch-yet' || data.status === 'nodiff' ) {
-                ubolog(`Diff updater: skip update of ${data.name} using ${data.patchPath}, reason: ${data.status}`);
-                assetCacheSetDetails(data.name, {
+                ubolog(`Diff updater: skip update of ${data.name} using ${data.patchPath}\n\treason: ${data.status}`);
+                const assetKey = assetKeyFromDiffName(data.name);
+                assetCacheSetDetails(assetKey, {
                     writeTime: data.lastModified || 0
                 });
                 vAPI.messaging.broadcast({
                     what: 'assetUpdated',
-                    key: data.name,
+                    key: assetKey,
                     cached: true,
                 });
             } else {